requirements.txt
streamlit
pandas
numpy
openai
python-dotenv
scikit-learn
faiss-cpu
